Split LSP lifecycle handlers into server_lifecycle.go
Move handleInitialize, handleInitialized, and registerWatchers out of
server.go into server_lifecycle.go. Pure file-split refactor — no
behavior change. server.go: 2130 -> 2008 lines.

Plan 203: start (status 🔲 -> 🔳).
